+++ src/main/java/constants/Constants.java
@@ -4,6 +4,4 @@ package constants;
 public class Constants {
     public static final String BASE_URI = "https://stellarburgers.nomoreparties.site/";
     public static final String ROOT = "/api/auth/";
-    public static final String EMAIL = "milkka2111@gmail.com";
-    public static final String PASSWORD = "qwertyi";
 }
