--- src/test/java/MoveThroughSection/MoveMainPageTest.java
@@ -1,47 +1,0 @@
-package MoveThroughSection;
-
-import io.qameta.allure.junit4.DisplayName;
-import org.junit.Test;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import pageObject.LoginPage;
-import pageObject.MainPage;
-import pageObject.ProfilePage;
-
-import java.time.Duration;
-
-import static pageObject.MainPage.ConstructorElement;
-
-public class MoveMainPageTest extends base.BaseTest {
-    @Test
-    @DisplayName("Проверь переход из личного кабинета в конструктор по клику на логотип Stellar Burgers")
-    public void MoveMainPageWithLogoButton() {
-        MainPage mainPage = new MainPage(driver);
-        ProfilePage profilePage = new ProfilePage(driver);
-        LoginPage loginPage = new LoginPage(driver);
-
-        mainPage.clickAccountButton();
-        loginPage.enterEmailAndPassword();
-        loginPage.clickSignInButton();
-        mainPage.clickAccountButton();
-        new WebDriverWait(driver, Duration.ofSeconds(5)).until(ExpectedConditions.elementToBeClickable(ConstructorElement));
-        profilePage.clickLogoButton();
-        mainPage.checkOrderButton();
-    }
-
-    @Test
-    @DisplayName("Проверь переход из личного кабинета в конструктор по клику на «Конструктор»")
-    public void MoveMainPageWithConstructorButton() {
-        MainPage mainPage = new MainPage(driver);
-        ProfilePage profilePage = new ProfilePage(driver);
-        LoginPage loginPage = new LoginPage(driver);
-
-        mainPage.clickAccountButton();
-        loginPage.enterEmailAndPassword();
-        loginPage.clickSignInButton();
-        mainPage.clickAccountButton();
-        new WebDriverWait(driver, Duration.ofSeconds(5)).until(ExpectedConditions.elementToBeClickable(ConstructorElement));
-        profilePage.clickConstructorButton();
-        mainPage.checkOrderButton();
-    }
-}
