+++ src/test/java/base/BaseTest.java
@@ -5,7 +5,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.openqa.selenium.WebDriver;
 
-import pageObject.MainPage;
+import pageobject.MainPage;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -14,7 +14,7 @@ public class BaseTest {
     public WebDriver driver;
 
     @Before
-    public void setUpChrome () throws IOException {
+    public void setUpBrowser () throws IOException {
         Browser browser = new Browser();
         driver = browser.initDriver();
         driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
