+++ src/test/java/login/LoginTest.java
@@ -1,21 +1,34 @@
 package login;
 
+import info.UserClient;
+import info.UserCreate;
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
-import pageObject.ForgotPasswordPage;
-import pageObject.LoginPage;
-import pageObject.MainPage;
-import pageObject.RegisterPage;
+import pageobject.ForgotPasswordPage;
+import pageobject.LoginPage;
+import pageobject.MainPage;
+import pageobject.RegisterPage;
 
 public class LoginTest extends base.BaseTest {
+    private UserClient userClient;
+    private UserCreate testUser;
+
+    @Before
+    public void setUp() {
+        userClient = new UserClient(); // Инициализация клиента для работы с API
+        testUser = new UserCreate("milkka2111", "qwertyi", "milkka2111@gmail.com"); // Создание нового пользователя
+        userClient.register(testUser); // Регистрация пользователя через API
+    }
+
     @Test
     @DisplayName("Проверь вход через кнопку «Личный кабинет»")
     public void loginThroughPersonalAccountButton() {
         MainPage mainPage = new MainPage(driver);
         LoginPage loginPage = new LoginPage(driver);
-
         mainPage.clickAccountButton();
-        loginPage.enterEmailAndPassword();
+        loginPage.enterEmailAndPassword(testUser.getEmail(), testUser.getPassword());
         loginPage.clickSignInButton();
         mainPage.checkOrderButton();
     }
@@ -27,14 +40,14 @@ public class LoginTest extends base.BaseTest {
         LoginPage loginPage = new LoginPage(driver);
 
         mainPage.clickSignInButton();
-        loginPage.enterEmailAndPassword();
+        loginPage.enterEmailAndPassword(testUser.getEmail(), testUser.getPassword());
         loginPage.clickSignInButton();
         mainPage.checkOrderButton();
     }
 
     @Test
     @DisplayName("Проверь вход через кнопку в форме восстановления пароля.")
-    public void LoginThroughTheButtonInThePasswordRecoveryForm() {
+    public void loginThroughTheButtonInThePasswordRecoveryForm() {
         MainPage mainPage = new MainPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         ForgotPasswordPage forgotPasswordPage = new ForgotPasswordPage(driver);
@@ -42,7 +55,7 @@ public class LoginTest extends base.BaseTest {
         mainPage.clickSignInButton();
         loginPage.clickRestorePasswordButton();
         forgotPasswordPage.clickSignInButton();
-        loginPage.enterEmailAndPassword();
+        loginPage.enterEmailAndPassword(testUser.getEmail(), testUser.getPassword());
         loginPage.clickSignInButton();
         mainPage.checkOrderButton();
     }
@@ -57,8 +70,13 @@ public class LoginTest extends base.BaseTest {
         mainPage.clickSignInButton();
         loginPage.clickRegisterButton();
         registerPage.clickSignInButton();
-        loginPage.enterEmailAndPassword();
+        loginPage.enterEmailAndPassword(testUser.getEmail(), testUser.getPassword());
         loginPage.clickSignInButton();
         mainPage.checkOrderButton();
+    }
+    @After
+    public void tearDown() {
+        userClient.deleteUser(testUser.getEmail());
+        driver.quit();
     }
 }
