+++ src/test/java/login/LogoutTest.java
@@ -1,31 +1,51 @@
 package login;
 
+import info.UserClient;
+import info.UserCreate;
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import pageObject.LoginPage;
-import pageObject.MainPage;
-import pageObject.ProfilePage;
+import pageobject.LoginPage;
+import pageobject.MainPage;
+import pageobject.ProfilePage;
 
 import java.time.Duration;
 
-import static pageObject.LoginPage.ConstructorElement;
+import static pageobject.LoginPage.constructorElement;
 
 public class LogoutTest extends base.BaseTest {
+
+    private UserClient userClient;
+    private UserCreate testUser;
+
+    @Before
+    public void setUp() {
+        userClient = new UserClient(); // Инициализация клиента для работы с API
+        testUser = new UserCreate("milkka2111", "qwertyi", "milkka2111@gmail.com"); // Создание нового пользователя
+        userClient.register(testUser); // Регистрация пользователя через API
+    }
+
     @Test
     @DisplayName("Проверь выход по кнопке «Выйти» в личном кабинете.")
-    public void Logout() {
+    public void logout() {
         MainPage mainPage = new MainPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         ProfilePage profilePage = new ProfilePage(driver);
 
         mainPage.clickAccountButton();
-        loginPage.enterEmailAndPassword();
+        loginPage.enterEmailAndPassword(testUser.getEmail(), testUser.getPassword());
         loginPage.clickSignInButton();
-        new WebDriverWait(driver, Duration.ofSeconds(5)).until(ExpectedConditions.elementToBeClickable(ConstructorElement));
+        new WebDriverWait(driver, Duration.ofSeconds(5)).until(ExpectedConditions.elementToBeClickable(constructorElement));
         mainPage.clickAccountButton();
         profilePage.clickLogoutButton();
         loginPage.checkRegistrationIsSuccessfully();
+    }
+    @After
+    public void tearDown() {
+        userClient.deleteUser(testUser.getEmail());
+        driver.quit();
     }
 }
