+++ src/test/resources/readme.md
@@ -1,0 +1,2 @@
+chrome driver: chromedriver-mac-arm64-131.0.6778.204
+yandex driver: yandexdriver-24.12.0.1772-mac
